Move ArcFollowCamera into followCamera.ts
